+++ src/components/private-route.tsx
@@ -5,7 +5,7 @@ type PrivateRouteProps = {
 };
 
 export default function PrivateRoute({children}: PrivateRouteProps): JSX.Element {
-  const hasAccess = false;
+  const hasAccess = true;
 
   return hasAccess ? children : <Navigate to={'/login'} />;
 }
